+++ hutool-core/src/test/java/cn/hutool/core/exceptions/ExceptionUtilTest.java
@@ -13,25 +13,25 @@ import java.io.IOException;
  */
 public class ExceptionUtilTest {
 
-    @Test
-    public void wrapTest() {
-        IORuntimeException e = ExceptionUtil.wrap(new IOException(), IORuntimeException.class);
-        Assert.assertNotNull(e);
-    }
+	@Test
+	public void wrapTest() {
+		IORuntimeException e = ExceptionUtil.wrap(new IOException(), IORuntimeException.class);
+		Assert.assertNotNull(e);
+	}
 
-    @Test
-    public void getRootTest() {
-        //查找入口方法
-        StackTraceElement ele = ExceptionUtil.getRootStackElement();
-        Assert.assertEquals("main", ele.getMethodName());
-    }
+	@Test
+	public void getRootTest() {
+		// 查找入口方法
+		StackTraceElement ele = ExceptionUtil.getRootStackElement();
+		Assert.assertEquals("main", ele.getMethodName());
+	}
 
-    @Test
-    public void convertTest() {
-//        RuntimeException e = new RuntimeException();
-        IOException ioException = new IOException();
-        IllegalArgumentException argumentException = new IllegalArgumentException(ioException);
-        IOException ioException1 = ExceptionUtil.convertFromOrSuppressedThrowable(argumentException, IOException.class, true);
-        Assert.assertNotNull(ioException1);
-    }
+	@Test
+	public void convertTest() {
+		// RuntimeException e = new RuntimeException();
+		IOException ioException = new IOException();
+		IllegalArgumentException argumentException = new IllegalArgumentException(ioException);
+		IOException ioException1 = ExceptionUtil.convertFromOrSuppressedThrowable(argumentException, IOException.class, true);
+		Assert.assertNotNull(ioException1);
+	}
 }